+++ examples/voice-assistant/minimal_assistant.py
@@ -1,12 +1,25 @@
 import asyncio
+import logging
 
 from dotenv import load_dotenv
 from livekit import rtc
-from livekit.agents import AutoSubscribe, JobContext, WorkerOptions, cli, llm
+from livekit.agents import (
+    AutoSubscribe,
+    JobContext,
+    JobProcess,
+    WorkerOptions,
+    cli,
+    llm,
+)
 from livekit.agents.voice_assistant import VoiceAssistant
 from livekit.plugins import deepgram, openai, silero
 
 load_dotenv()
+logger = logging.getLogger("voice-assistant")
+
+
+def prewarm(proc: JobProcess):
+    proc.userdata["vad"] = silero.VAD.load()
 
 
 async def entrypoint(ctx: JobContext):
@@ -18,16 +31,27 @@ async def entrypoint(ctx: JobContext):
         ),
     )
 
+    logger.info(f"connecting to room {ctx.room.name}")
     await ctx.connect(auto_subscribe=AutoSubscribe.AUDIO_ONLY)
 
+    # wait for the first participant to connect
+    participant = await ctx.wait_for_participant()
+    logger.info(f"starting voice assistant for participant {participant.identity}")
+
+    dg_model = "nova-2-general"
+    if participant.kind == rtc.ParticipantKind.PARTICIPANT_KIND_SIP:
+        # use a model optimized for telephony
+        dg_model = "nova-2-phonecall"
+
     assistant = VoiceAssistant(
-        vad=silero.VAD.load(),
-        stt=deepgram.STT(),
+        vad=ctx.proc.userdata["vad"],
+        stt=deepgram.STT(model=dg_model),
         llm=openai.LLM(),
         tts=openai.TTS(),
         chat_ctx=initial_ctx,
     )
-    assistant.start(ctx.room)
+
+    assistant.start(ctx.room, participant)
 
     # listen to incoming chat messages, only required if you'd like the agent to
     # answer incoming messages from Chat
@@ -48,4 +72,4 @@ async def entrypoint(ctx: JobContext):
 
 
 if __name__ == "__main__":
-    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint))
+    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint, prewarm_fnc=prewarm))