+++ livekit-plugins/livekit-plugins-openai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # livekit-plugins-openai
 
+## 0.8.2
+
+### Patch Changes
+
+- Add deepseek LLMs at OpenAI plugin - [#714](https://github.com/livekit/agents/pull/714) ([@lenage](https://github.com/lenage))
+
+- skip processing of choice.delta when it is None - [#705](https://github.com/livekit/agents/pull/705) ([@theomonnom](https://github.com/theomonnom))
+
 ## 0.8.1
 
 ### Patch Changes